requirements.txt
Django==1.7.1
# GDAL requires some special treatment, listed below for version info.
#GDAL==1.10.0 
argparse==1.2.1
django-flatblocks==0.8
django-geojson==2.6.0
django-jquery==1.9.1
django-leaflet==0.15.0
django-pdb==0.4.0
django-pipeline==1.4.2
django-shapes==0.2.0
futures==2.2.0
lxml==3.4.1
psycopg2==2.5.4
six==1.8.0
wsgiref==0.1.2
